# src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"TaskPriority": {
				"enum": [
					"MUST_DO",
					"SHOULD_DO",
					"COULD_DO"
				],
				"type": "string"
			},
			"TaskStatus": {
				"enum": [
					"IN_PROGRESS",
					"COMPLETED"
				],
				"type": "string"
			},
			"TaskResponse": {
				"properties": {
					"taskId": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"priority": {
						"$ref": "#/components/schemas/TaskPriority"
					},
					"memo": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"$ref": "#/components/schemas/TaskStatus"
					},
					"taskDate": {
						"type": "string",
						"format": "date-time"
					},
					"completedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"userId": {
						"type": "string"
					},
					"tagId": {
						"type": "string"
					},
					"tag": {
						"properties": {
							"projectName": {
								"type": "string",
								"nullable": true
							},
							"color": {
								"type": "string",
								"nullable": true
							},
							"tagName": {
								"type": "string"
							},
							"tagId": {
								"type": "string"
							}
						},
						"required": [
							"projectName",
							"color",
							"tagName",
							"tagId"
						],
						"type": "object"
					}
				},
				"required": [
					"taskId",
					"title",
					"priority",
					"memo",
					"status",
					"taskDate",
					"completedAt",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"userId",
					"tagId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TaskResponse-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/TaskResponse"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"success",
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TaskResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"allOf": [
							{
								"$ref": "#/components/schemas/TaskResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTaskRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"priority": {
						"$ref": "#/components/schemas/TaskPriority"
					},
					"taskDate": {
						"type": "string"
					},
					"tagId": {
						"type": "string"
					},
					"memo": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"priority",
					"taskDate",
					"tagId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateTaskRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"priority": {
						"$ref": "#/components/schemas/TaskPriority"
					},
					"status": {
						"$ref": "#/components/schemas/TaskStatus"
					},
					"taskDate": {
						"type": "string"
					},
					"tagId": {
						"type": "string"
					},
					"memo": {
						"type": "string",
						"nullable": true
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"anyOf": [],
						"nullable": true
					}
				},
				"required": [
					"success",
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TagResponse": {
				"properties": {
					"tagId": {
						"type": "string"
					},
					"tagName": {
						"type": "string"
					},
					"color": {
						"type": "string",
						"nullable": true
					},
					"projectName": {
						"type": "string",
						"nullable": true
					},
					"projectPurpose": {
						"type": "string",
						"nullable": true
					},
					"expectedOutcome": {
						"type": "string",
						"nullable": true
					},
					"expectedStartDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"expectedEndDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"kpis": {},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"userId": {
						"type": "string"
					}
				},
				"required": [
					"tagId",
					"tagName",
					"color",
					"projectName",
					"projectPurpose",
					"expectedOutcome",
					"expectedStartDate",
					"expectedEndDate",
					"kpis",
					"createdAt",
					"updatedAt",
					"deletedAt",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TagResponse-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/TagResponse"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"success",
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TagResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"allOf": [
							{
								"$ref": "#/components/schemas/TagResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTagRequest": {
				"properties": {
					"tagName": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"projectName": {
						"type": "string"
					},
					"projectPurpose": {
						"type": "string"
					},
					"expectedOutcome": {
						"type": "string"
					},
					"expectedStartDate": {
						"type": "string"
					},
					"expectedEndDate": {
						"type": "string"
					},
					"kpis": {}
				},
				"required": [
					"tagName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateTagRequest": {
				"properties": {
					"tagName": {
						"type": "string"
					},
					"color": {
						"type": "string"
					},
					"projectName": {
						"type": "string"
					},
					"projectPurpose": {
						"type": "string"
					},
					"expectedOutcome": {
						"type": "string"
					},
					"expectedStartDate": {
						"type": "string"
					},
					"expectedEndDate": {
						"type": "string"
					},
					"kpis": {}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "wordy_be",
		"version": "1.0.0",
		"description": "Wordy 백엔드 리포지토리입니다.",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/tasks": {
			"get": {
				"operationId": "GetTasksByDate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TaskResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "date",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TaskResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTaskRequest"
							}
						}
					}
				}
			}
		},
		"/tasks/{taskId}": {
			"get": {
				"operationId": "GetTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TaskResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TaskResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTaskRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTask",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Tasks"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "taskId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/tags": {
			"get": {
				"operationId": "GetTags",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TagResponse-Array_"
								}
							}
						}
					}
				},
				"summary": "태그 목록 조회",
				"tags": [
					"Tags"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateTag",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TagResponse_"
								}
							}
						}
					}
				},
				"summary": "태그 생성",
				"tags": [
					"Tags"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTagRequest"
							}
						}
					}
				}
			}
		},
		"/tags/{tagId}": {
			"get": {
				"operationId": "GetTag",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TagResponse_"
								}
							}
						}
					}
				},
				"summary": "태그 상세 조회",
				"tags": [
					"Tags"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "tagId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateTag",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TagResponse_"
								}
							}
						}
					}
				},
				"summary": "태그 수정",
				"tags": [
					"Tags"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "tagId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTagRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTag",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"summary": "태그 삭제",
				"tags": [
					"Tags"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "tagId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}